// front/src/api/models/ContinuousMeta.ts
/* tslint:disable */
/* eslint-disable */
/**
 * demo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ContinuousMeta
 */
export interface ContinuousMeta {
    /**
     * 
     * @type {string}
     * @memberof ContinuousMeta
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof ContinuousMeta
     */
    min: number;
    /**
     * 
     * @type {number}
     * @memberof ContinuousMeta
     */
    max: number;
}

export function ContinuousMetaFromJSON(json: any): ContinuousMeta {
    return ContinuousMetaFromJSONTyped(json, false);
}

export function ContinuousMetaFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContinuousMeta {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'min': json['min'],
        'max': json['max'],
    };
}

export function ContinuousMetaToJSON(value?: ContinuousMeta | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'min': value.min,
        'max': value.max,
    };
}